Establish the reference data set for the additional taxi-in/taxi-out method.

Establish the reference data set for the additional taxi-in/taxi-out method.

Usage

build_txxt_reference(refdf, variant = NULL, ref_period = NULL)

Arguments

refdf

data frame/tibble of taxi-times serving as reference data set

variant

of the determination of the reference times (i.e., GANP, PBWG)

ref_period

the reference period (currently: year)